fixed code block handling
